refactored for reusability

+++ src/RepoDownloader.py
@@ -1,9 +1,18 @@
 from github import Github
 from datetime import datetime
 import math
-import time, os, subprocess, shutil
+import time
+import os
+import subprocess
+import shutil
 
-def filterRepo(rootDir):
+GITHUB_USER_NAME = 'brokenquark'
+GITHUB_PASSWORD = '***'
+REPO_DOWNLOAD_DIRECTORY = '/home/brokenquark/Workspace/SLIC-Ansible/repo-github/'
+APPLY_FILTER = False
+
+
+def filterRepo(rootDir, applyFilter):
     paths = ''
     count = 0
     unnecessaryScriptCount = 0
@@ -20,7 +29,8 @@ def filterRepo(rootDir):
             if (not fileName.endswith('yml') and not fileName.endswith('yaml')) or 'test' in repo['path']:
                 unnecessaryScriptCount += 1
                 try:
-                    if '.git' not in repo['path']: os.remove(f'{dirName}/{fileName}')
+                    if '.git' not in repo['path']:
+                        os.remove(f'{dirName}/{fileName}')
                 except:
                     pass
             else:
@@ -28,7 +38,7 @@ def filterRepo(rootDir):
                 paths = paths + f'{dirName}/{fileName}\n'
 
     status = None
-    if 100 * (unnecessaryScriptCount / count) > 90:
+    if 100 * (unnecessaryScriptCount / count) > 90 and applyFilter == True:
         status = 'NOT_ENOUGH_PLAYBOOKS'
         try:
             shutil.rmtree(f'{rootDir}')
@@ -39,89 +49,56 @@ def filterRepo(rootDir):
         # file = open('ymlpaths2.txt', 'a')
         # file.write(paths)
         # file.close()
-    print(f'\t#########--{(100 - 100*(unnecessaryScriptCount/count)):.2f}--#########')
+    print(
+        f'\t#########--{(100 - 100*(unnecessaryScriptCount/count)):.2f}--#########')
     return [status, (100 - 100*(unnecessaryScriptCount/count))]
 
 
-def cloneRepo(owner, repoName):
-    g = Github('brokenquark','******')
-    # g = Github()
-
-    #repo = g.get_repo(f'{owner}/{repoName}')
+def cloneRepo(owner, repoName, applyFilter):
+    g = Github(GITHUB_USER_NAME, GITHUB_PASSWORD)
 
     status = None
 
     try:
         repo = g.get_repo(f'{owner}/{repoName}')
     except:
-        # print('not found')
         status = ['============NOT_FOUND============', '']
         return status
 
-    if not repo.fork:
+    if not repo.fork or applyFilter == False:
         contributors = repo.get_contributors()
         authors = []
         for item in contributors:
             authors.append(item.login)
         numberOfAuthors = len(authors)
 
-        if numberOfAuthors > 9:
+        if numberOfAuthors > 9 or applyFilter == False:
             commits = repo.get_commits()
             numberOfCommits = commits.totalCount
 
-            firstCommit = datetime.strptime(commits[0].raw_data['commit']['committer']['date'], '%Y-%m-%dT%H:%M:%SZ')
-            lastCommit = datetime.strptime(commits.reversed[0].raw_data['commit']['committer']['date'], '%Y-%m-%dT%H:%M:%SZ')
+            firstCommit = datetime.strptime(
+                commits[0].raw_data['commit']['committer']['date'], '%Y-%m-%dT%H:%M:%SZ')
+            lastCommit = datetime.strptime(
+                commits.reversed[0].raw_data['commit']['committer']['date'], '%Y-%m-%dT%H:%M:%SZ')
 
-            if math.fabs((lastCommit - firstCommit).days) > 0:
-                commitsPerMonth = math.ceil(math.fabs(numberOfCommits / ((lastCommit - firstCommit).days/30)))
+            if math.fabs((lastCommit - firstCommit).days) > 0 or applyFilter == False:
+                commitsPerMonth = math.ceil(
+                    math.fabs(numberOfCommits / ((lastCommit - firstCommit).days/30)))
             else:
-                # print('very short timed repository')
-                status = ['NOT_ENOUGH_COMMITS','']
+                status = ['NOT_ENOUGH_COMMITS', '']
                 return status
 
             dir = f'{owner}@{repoName}'
-            if commitsPerMonth >= 2:
+            if commitsPerMonth >= 2 or applyFilter == False:
                 gitUrl = f'https://github.com/{owner}/{repoName}'
-                dir = f'/home/brokenquark/repo-ansi/{dir}/'
+                dir = f'{REPO_DOWNLOAD_DIRECTORY}{dir}/'
                 subprocess.call(f'git clone {gitUrl} {dir}', shell=True)
-                status = filterRepo(dir)
+                status = filterRepo(dir, applyFilter)
             else:
-                # print('not enough commits')
-                status = ['NOT_ENOUGH_COMMITS','']
+                status = ['NOT_ENOUGH_COMMITS', '']
         else:
-            # print('not enough authors')
-            status = ['NOT_ENOUGH_DEVELOPERS','']
+            status = ['NOT_ENOUGH_DEVELOPERS', '']
     else:
-        # print('is forked')
-        status = ['IS_FORKED','']
+        status = ['IS_FORKED', '']
     return status
-
-repoList = []
-
-file = open('repoList/repo-ansible.csv', 'r')
-for line in file:
-    repoList.append(f'{line.split(",")[0]}/{line.split(",")[1]}')
-
-filteredRepoList = set(repoList)
-filteredRepoList = list(filteredRepoList)
-filteredRepoList.sort()
-print(len(filteredRepoList))
-
-for i in range(0, 1):
-    # if i%1000==0:
-    #     print('waiting...')
-    #     time.sleep(60)
-
-    owner = filteredRepoList[i].split('/')[0]
-    repo= filteredRepoList[i].split('/')[1]
-    # print(f'#{i} : {owner}@{repo}')
-    try:
-        status = cloneRepo(owner, repo)
-        print(f'{i} | {owner}@{repo} | {status}')
-        file = open('repo-stat/ansi-repo-stat.csv', 'a')
-        file.write(f'{i},{owner}@{repo},{status[0]},{status[1]},\n')
-        file.close()
-    except:
-        print(f'terminating at {i}')
-        break
